Expose edit notices so VisualEditor can see them
This exposes the notices generated by FlaggablePageView::addToEditView()
through the GetEditNotices hook, so VisualEditor can see them.
However, for EditPage we display the notices differently (with inline
collapsible diffs and log excerpts) and wrap them in a table, so we only
add notices to getEditNotices() if the caller isn't EditPage. The hack we
use to detect this is checking if addToEditView() has already been called.

This change is pretty hacky but I don't see how this could be done much
better given the strong reliance on EditPage in the surrounding code.

Change-Id: I78815cf7d308a42692d8856f7de4acc570ecfaeb