Make wpReviewthis checkbox actually work
WebRequest::getBool('fieldname', true) always returns true, because
it is impossible to distinguish between no form being submitted and form
being submitted with the checkbox unchecked.

This also has effect of making the checkbox in question unchecked by
default.  This behavior is in fact desirable, because otherwise the
users could mark potentially bad changes as reviewed without realizing
that.

Change-Id: Ib57455d4275174dc46434b3ae02fdffefc5aaa35