Autopromote: Remove superflous condition in APCOND_EDITCOUNT
This removes an extra condition in the core APCOND_EDITCOUNT autopromote
criteria. Although it is stated that we can exclude the last X days from
the edit counts, this is not possible for core APCOND_EDITCOUNT as seen
in Autopromote.php. This only works for FlaggedRevs' autopromote
conditions APCOND_FR_CONTENTEDITCOUNT and APCOND_FR_CHECKEDEDITCOUNT.

Change-Id: Ie3d9af5f97d03fb2d3a7d913f3a9210cc7dbd0b4