Apply the nomobile class
The ext.flaggedRevs.basic is not currently targeted at mobile and looking
at it, there is no need for it to be.

Unfortunately this means the banner in FlaggablePageview is visible in mobile
browsers taking up real estate.

This generic solution allows the hiding of this banner at mobile resolutions

Bug: T119344
Change-Id: I8c718394f870812ec946dffb2b89474757c64b08