Pass new stable revision to review log for dis-approvals
Currently, the old stable revision is passed to the review log when a
revision is unreviewed, same as when a revision is reviewed. But then
the changes that were unreviewed are never shown in the log, since the
old stable revision is the same as the unreviewed revision, or more recent.
Instead it should be the new stable revision that gets pased. This way,
the log will show which changes were unreviewed. The fix doesn't work for
old revisions that get unreviewed, but in that case it would be a pain to
know which prior revision to base the diff on.

Change-Id: Ib3d709679396ace8b794648c827685b209a7e5f7