Show log excerpt by default in review form
This shows the stable log excerpt by default in the review form because
it can contain very useful information to reviewers. On enwiki, it is
required by policy to check the reason for pending changes protection.
Thus showing is by default is sensible.
The log excerpt shown when editing remains hidden by default.

Change-Id: I626e8c251f5c06acec93ce72d7cebd77901cbd12