Remove reliance on baseRevId and altBaseRevId for FlaggablePageView
This checks if there are pending changes instead of relying on
baseRevIds / altBaseRevIds to predict whether the edit will be
autoreviewed. This information is already fetched for the
"review this edit" form for reviewers, and it won't
significantly affect response time for non-reviewers.
They were already no longer needed to check if the edit should
be autoreviewed post-save, they are now completely unused, so
are removed altogether.

Change-Id: Ife46a74b0652c39cf65b70b9777853f583fee70c