Put fileSearchOptions() modification back to where it belongs
The fileSearchOptions() modification that happens when the stable
versions of files is retrieved during autoreview is put back to
where it belongs, i.e. inside the closing braces above it.

This was accidentally modified during the refactoring of the class
in I31d0081ebda3c521635dd3abb40ec4dd1aad57c2.

This didn't result in any visible change to end-users, the only
effect was to run queries that were only necessary to be run on
projects using full FlaggedRevs like dewiki, which however were
completely unecessary on projects like enwiki.

Change-Id: I2bd8eccfdc7ac1e74b048ad9a272c1952d454dec