Remove usage of EnqueueJob
Since the flaggedrev_CacheUpdate job was moved to the
kafka JobQueue the usage of EnqueueJob became unnesessary.
Initially it was designed to post cross-DC jobs, but in the
new infrastructure it's not needed, cross-DC is handled by
the JobQueue itself.

Additionally, the Job contract requires the job to return
a boolean indicating if the job execution was successfull
or a job should be retried. Added return statements everywhere.

Bug: T181216
Change-Id: I79167ee190120b4ef9220ff906635003e7b2439e