FlaggablePageView: Let other editors know which revision of the page to edit
When the user is viewing the last stable revision of a page which has
newer unreviewed revisions, FlaggedRevs wants us to open the latest
(unreviewed) revision of the page for editing.

Provide a JS config variable 'wgFlaggedRevsEditLatestRevision' to
indicate this, because this behavior depends on a lot of things that
are difficult to exactly replicate in other code.

Bug: T165283
Change-Id: I4c9804fe2c4924e28770807881379ddca4fd8b76