Replace old and busted hook with the new hotness of a callback
FlaggedRevs overrides the revisions of templates used when parsing
a stable revision of a page. It will now do so by using the
ParserOptions::setCurrentRevisionCallback() method instead of
the BeforeParserFetchTemplateAndtitle hook.

This makes FlaggedRevs compatible with extensions that bypass
the hook, like TemplateStyles.

FlaggedRevs now requires at least MediaWiki 1.24 (no idea which
version it required before).

Bug: T190945
Change-Id: I8101367ca190e8c33f78bffbcf342e8013b0c9e7