Fix compat with old local overrides for 'revreview-sadv-reviewing-c', 'revreview-sadv-reviewing-p'
If these messages were overridden locally on a wiki, the recent
changes from 0a2e9a12244a921b5f6e048af8a1f280fdd5621f would cause a
bunch of JavaScript code to be inserted into the page instead of the
"advertise" link.

This change fixes it by explicitly building a clickable `<a>` element
and passing it as the parameter, which is compatible with both old and
new message contents.

To do this we need to restore the 'revreview-adv-start-link' message,
the translations should be automatically restored from Translatewiki.

Bug: T196104
Change-Id: Ibbd15f1811bca5e649ef5acebd8a0d543e887a2e