FlaggedRevision: handle mismatching page ID.
Apparently, flaggedpages.fp_stable sometimes points to a revision that
does not belong to the page indicated by flaggedpages.fp_page_id. This
leads to a fatal error.

This patch detects this situation and prevents a fatal error in
FaggedRevision. It also adjusts the code in FlaggableWikiPage to make it
less prone to inconsistencies. Whether this is going to fix the root
cause of the problem is unknown.

Bug: T246720
Change-Id: I2826c6b448ece5b1d4b4c7cb0a7f4ce01e6e22cc