Remove a ton of meaningless "return true"
It seems there is (was?) some kind of convention in place to always
return true from just every piece of code, even if the return value
is never going to be used, and even if there is no way it can ever
be false. This is already super inconsistent. A lot of code doesn't
do it. Some methods currently return true|null, because one
"return true" was forgotten.

Let's please get rid of this visual distraction.

Many of the "return true" are in hook handlers. These default to
true anyway for many, many years now. Returning null/void is the
same as returning true.

Note: There are a few more pointless "return true". However, they
are in more complicated code. I would like to continue cleaning
this up in later patches.

Change-Id: Ida3a8ffaa893a805bcf924d7a50f03db9d1b4855