Remove not needed array initializations
PHP will happily create nested arrays for us whenever we set a
value like $a['first']['second'] = true. The only thing that
must exist before is the very first array level $a = [].

I know there is no strict rule for one or the other style.
Having these extra initializations does have an advantage.
However: I would like to argue that the repeated (3 times!)
code is more dangerous to have than not to have. It's so easy
to have an error in there, and almost impossible to spot.

Change-Id: I81637abad607a10b84441160cf9e150c15e0b545