Don't declare properties "protected by default" when not needed
There is only one reason to make something protected: if there
is a plan to create subclasses, and if the subclass should have
access to a property. Neither is the case here. I would even go
so far and argue that subclassing random classes needs to be
forbidden, unless a class is specifically designed to allow this.
The "protected" visibility level creates the false impression
that subclassing would be ok, when it isn't.

Validating this patch is rather trivial. We don't even need to
check these properties individually. All we need to do is to
check if a subclass exists. E.g. via
https://codesearch.wmcloud.org/search/?q=extends.*FlaggablePageView
If no subclass exists, "protected" is meaningless and can
blindly be replaced with "private".

Note this patch is exclusively about protected properties,
nothing else.
* No public properties are touched, even if some could be made
  private as well. This is for a separate patch.
* No methods are touched. This is as well for another patch.

Change-Id: Ia1719e39c752f987b983ae506ad5369607566688