Inline pointless "table" parameter in FRExtraCacheUpdateJob
This never changes, and never even can change. The query requests
a field "ftr_from", and that doesn't exist anywhere, except in
this table.

Even worse: There are 3 places in this codebase that create a
FRExtraCacheUpdateJob, but only 1 populates the "table" parameter.
Luckily this is not an actual bug because this is not really 1 job
but 3 different ones. Each call uses a different "type". Only 1
requires the "table". Still, what's the point? Let's just inline
this, please.

Change-Id: Ia7e1b5b092c14c56efc5559238dc0ecc782820be