Mark private methods a such, not "protected by default"
Same as Ia1719e3, but for methods.

There are only 2 reasons why a method needs to be protected:
* If there is a subclass that needs to call the method.
* When it's an override for a protected method in a base class.

Which means reviewing this is rather easy in most cases:
* Check if the class extends or implements anything. If not, then
  there is no base class. Look for traits as well.
* Check if a class does have subclasses. If not, no other code
  can call these protected methods.

Both is true for most of the classes touched in this patch.
Some extend a base class. In these cases I manually checked if a
method overrides something from the base class or not. IDEs (I
use PHPStorm) are really good making this easily visible.

Here is a quick way to find unknown subclasses:
https://codesearch.wmcloud.org/search/?q=(extends%7Cimplements%7Cuse)%5Cs%2B(ApiStabilizeGeneral%7CConfiguredPages%7CFlaggablePageView%7CFlaggableWikiPage%7CFlaggedRevision%7CFlaggedRevs%7CFlaggedRevsHooks%7CFlaggedRevsUIHooks%7CFRDependencyUpdate%7CFRExtraCacheUpdate%7CFRExtraCacheUpdateJob%7CFRInclusionCache%7CFRInclusionManager%7CFRPageConfig%7CFRUserActivity%7CFRUserCounters%7CPageStabilityForm%7CPendingChanges%7CProblemChanges%7CReviewedPages%7CReviewedVersions%7CRevisionReview%7CRevisionReviewFormUI%7CStabilization%7CStablePages%7CUnreviewedPages%7CValidationStatistics)
There is only 1 exception, which we can check like this:
https://codesearch.wmcloud.org/search/?q=updateWatchlist&repos=Extension:FlaggedRevs

Change-Id: I8337618911b41413b4a4bf048bcf7baef55d1f8b