Rewrite overly complicated (and wrong!) purge job
Let's say we want to purge the page IDs 1, 2, 3, and 501, and
the limit for one job is 2, the old code did that:
* Run a purge job with IDs from 1 to 2.
* Run another job for all IDs from 3 to 501.
… wait, what? So 4, 5, 6, …, 499, and 500 are all purged as
well? Why? This doesn't happen in the other code path, when
the number of pages to purge is below the limit.

My algorithm is meant to work like this:
* Find consecutive sequences of IDs.
* When there is a gap between two IDs, start another job.
* If one job grows to big, start another one.
* If we reach the end, make sure the last ID is in a job as
  well.

The only negative effect of this is when there are no
consecutive IDs in the array (e.g. 1, 3, 5, 7, 9, …). The code
will run a job for every single ID then. Is this worth doing
something about? Personally, I don't think so.

Change-Id: I90820eac9bd0b850ae6170d2c0b6c82d4003d1ef