Move defaults in FlaggableWikiPage::loadPageData() up
This should not make any difference, except: In case a wiki
is not properly configured yet (i.e. update.php was not run),
the database access in the row $this->pageDataFromTitle( … )
will throw an exception. Nothing will be initialized, but
$this->mDataLoaded is already true. This causes many more
confusing errors on top of the exception.

With this change, only the exception is shown.

Change-Id: Ied47c44043b7142fa941b67c1b2e49ed189433d9