Guard more singleton() calls with globalArticleInstance() checks
Previously, only a handful of hook handlers checked if the
FlaggablePageView::globalArticleInstance() exists before accessing the
FlaggablePageView::singleton(). However, if the globalArticleInstance()
doesn’t exist – this is typically the case in unit tests, when the
global request context is not fully initialized – then the singleton can
be created, but will crash in load() as soon as you try to do something
with it. We should probably add this check to all hook handlers that
access the singleton.

The one hook handler that I left out is overrideRedirect() (registered
as handler for InitializeArticleMaybeRedirect). That function scares me
– it mixes an $fa for a given title with a $view for the global request
context title – and I’d rather not touch it unless I have to.

Bug: T269608
Change-Id: I03aa1ac6ed20c81cf131bb76dd817ad8cf34d215