Clean up two barely used frontend features
The first one gives a background image on a non-low profile wiki (two
wikis, including Hungarian Wikipedia) and when a page has a pending edit
and you're logged out and seeing the old version instead and you hover
on the eye icon. You'd see this background which actually makes it
harder to read (violating WCAG), without it it just looks fine.

The second one is even worse. The wiki need to be non-simple UI (only
three wikis), and has to have more than two levels (only English
Wikibooks), you need to have reviewer right, check out an old stable revision
(with oldid=) and then it's hidden under "display: none" on the message
box, you can inspect element it or with javascript show that +/- which
by clicking shows a bar of progress.

Bug: T277883
Change-Id: Ia662391b3e97963782abf7190e91ad6c54ddebc8