Wrap most of functionalities depending on protect mode in a condition
If you set protect mode on (aka simple mode), dimensions and tags won't
be filled so you'll have empty array and it just ignores most of the
code implicitly. Changing the code to have only one dimension broke this
assumption.

This patch is a follow up to I594f0c503602

Bug: T278478
Change-Id: I36d160d9617956523afa24df04504b67b4820cee