Make FlaggedRevsTest more robust, part 2
This test fails in the BlueSpiceFlaggedRevsConnector extension.
I think I understand why:

The BlueSpice extension contains a large wgExtensionFunctions
function that manipulates the global that is relevant for this
test. While the test tries to reset the global, this doesn't
have an effect. Some unrelated hook initialized the FlaggedRevs
class already.

The getLevels() feature also depends on another config which
should be forced to  match what the test expects.

Change-Id: I070f97eb199ac421f7f76d4d600c65354fffb199