Drop using ft_title and ft_namespace
flaggedtemplates mirrors templatelinks table (by picking up its worst parts),
for templatelinks it makes sense to keep title instead of page_id or rev_id
but for flaggedtemplates, we are keeping track of stable revisions
of transcluded templates.
That concept is useless here as a non-existent template doesn't have a stable version.

This is adding massive amount of useless data to wikis such as arwiki
and dewiki.

Bug: T296380
Change-Id: If31967c20ee8b6e1310f36e66d35760ef07a0742