Replace wfQueriesMustScale() and generalQueryOK() with MiserMode
The wfQueriesMustScale() function returned true for MiserMode, and for
non-miser wikis above a certain size. By reducing the check to just
MiserMode, this should functionally be safe in so far that it will not
disable anything or act more conservative than before in a way that may
diminish functionality. On the other hand, if this extension is enabled
on WMF prod wikis that are both non-miser and above the core threshold,
it's worth checking that the live queries perform fine.

The generalQueryOK() method did the opposite, when it found the wiki
is in miser mode (or the broader "must scale" mode), it would sometimes
still allow live queries if there is only a small fraction of pages
with FlaggedRevs enabled. The feaeture will now no longer vary on that,
meaning that if there are wikis with miser mode enabled and only a
small fraction of FlaggedRevs-enabled pages, it might be that it
currently enjoys live queries but would no longer.

If that is a problem, then we may want to remove the non-live mode
entirely if WMF doesn't need it, and maybe remove a simple config
flag for third parties if need the cached mode (e.g. if they use it
on all pages or something like that).

Change-Id: Ie3d8ab35c978458ee76663e3f0a38a4c269a6d5a