Remove db->freeResult()
None of these variables have been needed afterwards and they don't take
any space and php has a garbage collector (not to mention in most of
these codes, it's in a loop so the value gets reassigned two cycles
later).

The function is also deprecated and adds unneeded complexity. It's
YAGNI.

Bug: T286694
Change-Id: I88cd4954428716fe3e2254978bb3ca504de67c3c