frontend: Remove static "once" check from injectStyleAndJS()
This was added in 2010 with commit 76bdb6b4c9cdc65 (SVN r78163),
at a time when this was still adding raw script and stylesheet tags
to the HTML output.

Now that it is using ResourceLoader, this is okay to (sometimes) be
called twice on the same page. This is called once via
onBeforePageDisplay and sometimes (also) via the
onDifferenceEngineViewHeader hook.

Change-Id: I78215b74d01dd3c70723d18f5fe52665c711cd2a