Add APCu cache in FlaggableWikiPage::pageData()
This is causing an unconditional db query in every page view of a
flaggable page which in most wikis are popular pages.

At first, I wanted to to make it use WAN cache but invalidating this
after review or edit can get really complicated. This is to handle
spikes in page view, i.e. Michael Jackson effect.

Bug: T302880
Change-Id: Iaf93206201155d03ce3ad728864ad4ff080514d8