Simplify array_filter() calls
Without the strlen() callback the array_filter() will behave a little
different. Now it removes any "falsy" value including the string "0".
I want to argue that this is fine in these contexts, even desirable.
A "0" can never be valid in these contexts exacty because it is easy
to confuse for a false.

Change-Id: Ia2e1914e1dabb1d08b5056063060f93523e8ee4d