Inline code reading from $wgFlaggedRevsAutoReview
This is so trivial, let's just inline it. No need to cache it via
self::load().

The logic is intentionally simplified and the is_int() check removed.
Almost all invalid stuff will result in 0 anyway, e.g. when the setting
is null. The behavior is the same as before. And when stuff can be
turned into an integer, why not use it?

This is split from I021d8ab to make it easier to review and pin the
CI failure down.

Change-Id: I25e3131993d42ea7e4463469fadb583ade919cd1