Inline trivial FlaggedRevs::binaryFlagging() logic
I use reset() to extract the first (and only) array element. I don't
care about the length of the array or the key. This is all already
checked in the load() method above.

This is split from I021d8ab to make it easier to review and pin down
the CI failure.

Change-Id: I299d404469e9dd325c313a8ce8cc36baa14e6901