Fix incomplete FlaggedRevs::binaryFlagging() implementation
Introduced in I299d404. While the logic how to decide if we are in
"binary flagging" mode is correct, the old code that was doing this
was never called in "useOnlyIfProtected" mode. In this mode the
"binary flagging" flag was always true. This doesn't make a lot of
sense in my opinion, but is apparently critical.

Bug: T307972
Change-Id: I5b9c65a2040032539da5f53bcf67607a6301cc27