UnreviewedPages: Fix "Show redirects" link
When making links, parameters set to the bool "false" value are unset
from the URL. However, the 'hideredirs' parameter defaults to being
enabled when unset, so the link had the opposite effect than intended.

Specify the values explicitly to ensure the correct behavior.

Follow-up to 7059d0c23a, which changed `1 - $this->hideRedirs` to
`!$this->hideRedirs, changing the result from int to bool.

Change-Id: I5b58110c2525313a715e8eef2533ca8a1a72db63