FlaggableWikiPage: Handle revisions with same timestamp for pending list
The query in FlaggableWikiPage::updatePendingList to get a edit after a
specific timestamp fails, when the next revision has the same timestamp
as the previous one. This can happen in tests.

This results in the query
INSERT INTO "unittest_flaggedpage_pending"
(fpp_page_id,fpp_quality,fpp_rev_id,fpp_pending_since) VALUES
(1,0,'1',0)
where the 0 for fpp_pending_since is invalid in postgres timestamp data
type and the query fails. it works on mysql/sqlite, but the 0 is
interpreted as "now" later on
Also check the result of the query as a second guard against unexpected
values in the database

Change-Id: I06a34f613e77eccd59e1f1aaa4a7a6e2126e01ee