pruneRevData: Make cleaning in larger batches
Currently, it cleans flaggedtemplates data in a more fragmented way, it
collects revs to be deleted on of every page and deletes them in those
batches. This is inefficient as most pages might have one or two
revision to be cleaned and might remove only ten rows and then sleeping
for a bit. That's why running this script in any wiki practicly never
finishes, I have been running this for half a year and it hasn't got to
the end yet.

Bug: T296380
Change-Id: I6bc1b89e744c704c2326787622eb3181dff6d27c