Changes list filter: don't add fields that are already in the query
fp_stable and fp_pending_since are already added to the query
unconditionally by the ChangesListSpecialPageQuery hook, so it's not
necessary to add them again. Adding them again causes a query error.

rc_namespace is unconditionally in the query already due to
RecentChange::getQueryInfo(). It didn't cause a query error because it
was filtered out by array_unique().

Bug: T316026
Change-Id: I4c20b74b1604e1a3bcb57d74dac160e7d95b972d