Split large if-else in FlaggedRevision constructor into two
This patch intentionally doesn't change any behavior. The effectively
only change is the strict type for the first $row argument. This
makes the "else-throw" obsolete.

Some stuff turns out to be unused now. I would like to continue
working on this in separate patches to make review as easy as
possible.

Change-Id: Ie6cac8cb0834eb459bb28eb433da1d91e7fa4cbc