Drop string→array→string conversion from FlaggedRevision
The reason this was done was an attempt to use the same constructor
for database rows (where this is indeed a "flattened" string) as well
as arrays. We separated these, see newFromRow() above. By making the
now simplified constructor accept tags as an array we can drop some
pointless flatten→unflatten conversions.

Note the database field is not nullable. The extra strval() never did
anything.

Change-Id: I4ab84bf48fcec69b45320d95fc7d8afdc080ceb9