SpecialContributions: Make it possible to see pending changes only
When $wgFlaggedRevsProtection is false (most flaggedrevs-enabled
wikis), this table is already joined for Special:Contributions
queries, so this should not have a significant effect
performance-wise. The filter is disabled on
$wgFlaggedRevsProtection wikis for now.

This will be helpful for GrowthExperiments, as we want
to allow mentors see pending edits by their mentees
(by linking to a filtered view of Special:Contributions).

Bug: T321445
Change-Id: Ibbb5e3b70280fa12f7084aca0dba378e386e1f35