RevisionReviewForm.php: Only check is the revision text is deleted
Before we were using getVisibility() to determine if a revision could be reviewed, which is causing a bug (T57349). To fix this, now it is only checked whether the revision text is visible before allowing it to be approved. This will ensure that revisions with a hidden author or summary can be approved.

Bug: T57349
Change-Id: I6b14a227c278ca3f89c75146d0d5e79867b3fbc2