Move special case out of RevisionReviewFormUI::ratingFormTags
There is only one situation where the ratingFormTags() method can
return an empty array of $labels. When this happens ratingInputs()
will return an empty string because the foreach loop at the bottom
creates an empty array of $items, which implodes to an empty string.

Extract this special case so it becomes much more obvious what's
going on here.

Change-Id: I9364d7cac87adc363a43690a3942044e26a10c0a