Reduce APCu cache use in FlaggableWikiPage::pageData()
This cache was added in change Iaf93206201 (commit c926d8c409), but ends
up caching a bit too aggressively: in an EntitySchema test that creates
two successive edits on a page, we noticed that FlaggedRevs can end up
resetting the title’s mLatestID to an outdated revision ID. Let’s not
use the cache if the database connection is for a primary and not a
replica database; that should ensure we load the latest data when the
method is called from the PageUpdater, without affecting the more common
read-only page view case in production. (On smaller setups with only a
primary database, this effectively disables the cache altogether,
because isReadOnly() will always return false in such a setup even if
the connection was obtained for DB_REPLICA.)

Note that it’s still possible for the cache to return outdated data on a
subsequent read; I don’t see a good way to avoid that at the moment.
This change is only a partial improvement, not a full fix for potential
issues caused by the cache. (However, the cache is also fairly important
for performance, so we can’t just remove it completely.)

Bug: T334464
Change-Id: I713431bfc2ddaa8193a20cbbbb08dd806ba772a7
(cherry picked from commit 5df767b4163a755ea6bea7d406d2a3242c40a194)