Drop all secondary caching of configuration parameters
This drops the last remaining secondary caches from the FlaggedRevs
base class:

1) Constructing the $dimensions array is extremely cheap and not worth
caching at all. It's extremely short, usually 3 elements.

2) Verifying the $reviewNamespaces is a little more effort because of
the services involved and because the array might be a bit longer.
Instead of storing a copy we now remember if we validated it already.

3) The extra processing on $restrictionLevels was weird. Why should it
contain duplicates? Isn't this a configuration error? Why should it
it contain an empty string that never does anything? Let's report this
as a configuration error instead.

Duplicates can never do anything bad.
* The array is used to verify if a value is present or not. Duplicates
  don't make any difference then.
* Duplicates might show up in a UI. This does (and should!) look weird
  but can't do anything bad. It just doesn't matter which one you
  pick.

Bug: T340004
Change-Id: I1591b95c2ccb5be1f5931cefbd5c02a0cc65b293