Fix ReviewHandlerTest reusing existing test page name
When called with no argument the page name is always "UTPage". I
suspect this causes conflicts with other tests that use the same
page name, but don't delete it when they are done.

The test cases in this file also conflict because of this. The last
test case tries to set the page content to "SecondEdit". But a
previous test already did that, which results in a null-edit and the
change time being null. This was never the purpose of this test.

Bug: T340004
Change-Id: I6c803f60896299146f065720ddab3831ad607bed
(cherry picked from commit 60199a1794b4c9b1fd0eac9f3551a24012da715c)