Remove unused FlaggedRevs::getLevels()
It was effectively only used in a single, maybe two places. Most code
doesn't use this specific way to encode level numbers as message keys.
We can inline this in the few places that need to know it.

getMaxLevel() is still fully covered with tests we just added via
I7f28421.

Change-Id: Ib2b910f300a63dd5172c50b89e9b0aab5bad5c6e