RevisionReviewFormUI: Remove some traces of the tiers
Use a single nullable integer instead of an array for the current flag
of the page; there won’t be more than one anyways. Add some methods in
backend/ to support this and deprecate the ones that (seem to) support
multiple tiers. (They don’t actually support multiple tiers anymore.)

Bug: T277883
Change-Id: I80faceb27277a7cb81d8a37c47b3e91c855a8ca4