Hide level zero in the revision review form
Reviewing a page with level zero is the same as unreviewing it – except
that it’s not even possible, the review button is disabled in this case
(and the backend rejects the request, should JavaScript be disabled).
Removing the input for level zero declutters the UI and at the same time
makes the code simpler (in case levels 0 and 1 are available, the
special-casing is now a hidden input rather than a checkbox – no need to
care about labels, Codex components, client-side handling etc.).

Since the client-side code had to be touched anyway to remove the code
special-casing the zeroth level radio buttons, I used the oppurtinity to
get rid of the inline JavaScript emitted by the PHP code, replacing it
with a `data-` attribute, and to use native JavaScript instead of jQuery
where jQuery doesn’t make the code simpler (and to remove code dealing
with checkboxes and selects, as checkboxes have been removed in this
commit, while selects didn’t exist previously either).

On the server side, I also removed a stray </span>.

Bug: T277883
Bug: T191156
Change-Id: I7cf1996e67dd6c6e0cb179eb7565c37e6eb0cf2c