Add convenient FlaggedRevs::isReviewNamespace() method
Turns out this concept (where the input is an integer number, e.g.
directly from a database column) is used a lot in this codebase.
I think it justifies a separate method.

This reduces the direct calls to FlaggedRevs::getReviewNamespaces()
from code that is actually not interested in the list, but only
needs a boolean.

Change-Id: I58fc467ef9710018ccdf032a5848abeaa471118f