Add convenient FlaggedRevs::getFirstReviewNamespace()
I'm not 100% sure if this justifies a separate method. The concept
appears a few times in this codebase.

This further reduces the number of direct calls to the
FlaggedRevs::getReviewNamespaces() method. Now it's only called in
places that actually need the full list.

Note this changes one place from `-1` to `NS_MAIN`. This is
intentional. This is most likely unreachable anyway. Even if it is
reachable it's fine to continue with the main namespace as a default
(as done in all other places). The performance argument (never more
than one namespace) is not violated.

Change-Id: I87fd0f74f56452b2b94a0752f0ce5817cf82a286