Move Media namespace special case back to inReviewNamespace
This was a mistake I made in I58fc467 just a few days ago. It's
probably not even reachable in production, but was not correct as
pointed out in a comment in I58fc467.

Also change the bigger test setup so the methods are actually marked
as being used by the test. I tried to be to clever before. The IDE
couldn't fully understand that.

Change-Id: I07f7a735bd36a91b524dfbeab6e804b0a1c97b7c