Load codex-styles only for reviewers
codex-styles is only used in the review form, which is only shown for
reviewers. Avoid loading the module for users without review right. This
could be further optimized (only load if the form will actually appear,
e.g. not when looking at page history; or evaluate whether the smaller
codex-search-styles module would be enough), but this fixes the most
performance-sensitive scenario of logged-out page views. Reviewers will
probably have the CSS in cache anyway, so further optimizations would
help little increasing the performance.

Also fix the comment: while thinking about doing further optimizations,
I realized that the JS module couldn’t be skipped solely based on the
presence of the review form, because it also affects the edit form (it
changes the Publish button to Save for edits that won’t be reviewed on
saving).

Bug: T350514
Change-Id: I1e45d0dacf58b2e03886d3b986309e13e754e324