Use latest stable revision of templates instead of tracking them
This effectively gets rid of flaggedtemplates and all that entails. We
do this because flaggedtemplates is extremely taxing on the
infrastructure and has caused outages. The users don't need the exact
revision id of the template used when it was reviewed, they can simply
get the latest reviewed/stable version of that template which is already
impelmented in FR as the fallback.

If inclusion setting has been set to FR_INCLUDES_STABLE,
FRInclusionManager::getStableTemplateVersion() gets triggered and as
such latest stable revision will be used.

Also completely get rid of showing the stable version in diff as it
doesn't serve much purpose and only adds a lot of code and db queries.

Bug: T359529
Bug: T277883
Change-Id: Icb60e30aedc148f2e6d4e29d762e01180eee1d42