Replace most of the more trivial Xml::… form builder methods
I'm intentionally leaving a few of the more complex methods behind,
e.g. Xml::fieldset and others that create two linked elements.

This also fixes an actual bug in the Stabilization class. The issue
was introduced in 2019 via Iac1246a. In that patch the local variable
$value was dropped and replaced with $option[1]. But one of the usages
of $value was forgotten.

Bug: T341776
Change-Id: I6907a1c3b096a634b7568ad77264d093c62fb7f3