ReviewHandler: declare all fields to be strings.
All parameters are received as strings, because review.js takes the
values from HTML form elements without understanding their types.
This is not a problem since we pass the request body to
RevisionReview::doReview(), which is designed to handle submit
data from an HTML form, which is all strings anyway.

Bug: T305973
Needed-By: I2d00d320f59652040ff914005aad58bf3e724689
Change-Id: I17f7d0b89cb584eef5a7203a8f26cb5b100e2abb