Use consistent capitalization in mw-fr-revisiontag
There is a chip that displays on the top right of pages.
To tell the viewer whether the page has pending changes
or not.

Before this patch, the two possible statuses are "Checked"
and "review pending changes".

After this patch, the two possible statuses are "Checked"
and "Review pending changes". This makes the capitalization
consistent.

Change-Id: I22866186ba6b59bf45a6fc3e31b5f15d561b9960