Special:PendingChanges: remove parentheses from diff sizes
The parentheses are probably just from recycling code from the
history page. They are not needed for our use case, so this patch
removes them, resolving the linked ticket.

This has a flash of unstyled content because it's in advanced.
I could move it to basic if this is a problem. But all the other
.mw-fr-pending-changes-table styling is in advanced, so I put
it there for now.

Also includes some minor refactoring: renaming a variable,
adding a line break.

Bug: T372187
Change-Id: I6ecad8d099d43a71e004a4af761f6c515579b4f5