FlaggablePageView: fix PHP type error
$this->article->getParserOutput only returns ParserOutput|false.
Fix showUnreviewedVersion()'s return type to match so that it
doesn't throw errors.

Union types aren't introduced until PHP 8.0. I think we're still
on PHP 7.4 in production, so I haven't elected to use a union
type in the PHP code in this patch, just in the docblock.

The function that calls showUnreviewedVersion(), setPageContent(),
expects bool|ParserOutput|null, so this shouldn't cause any problems
there.

Bug: T373113
Change-Id: I25e7a14506c95f004384bf2e0060fb73810a16a9