Add bundlesize test for modules loading on page load
To support us monitoring the amount of CSS and JS we ship to users
its imperative we start tracking the size of these modules statically
so we are aware when we increase the size of them.

Even though mediawiki.codex.messagebox.styles is defined in core,
this module is not loaded on page load by default and only loaded when
FlaggedRevisions is installed, so that is why this is defined here.

Bug: T373017
Change-Id: Iab4963144ee9883997bd2f219f3efa900e73df6c