build: Disable MediaWiki.WhiteSpace.SpaceBeforeSingleLineComment sniff
There are 40 files with 443 findings.

This extension is unmaintenanced and has many comments after code lines.
Sniffs listed in <exclude> get autofixed by libup and indicate that the
sniff should be fixed in the long term.
Disable the sniff with serverity = 0 to indicate that
this sniff does not need fixing in this extension.

Change-Id: I81f05e8c0c8e8dfe6aca50d5d08d732964a96c12