PendingChanges: Don't count pages that aren't in a FlaggedRevs namespace
At the moment, if a wiki originally has FlaggedRevs enabled for a given
namespace, but then removes that namespace from
$wgFlaggedRevsNamespaces, any previously-reviewed pages in that
namespace that (when the namespace was removed from that config
variable) _did_ have unreviewed changes will still be included in the
count on the top/bottom rows of the table at Special:PendingChanges.

In addition, if a page that's currently pending review is moved from a
FlaggedRevs-enabled namespace to a non-FlaggedRevs namespace, that page
will also currently still be included in those counts.

Pages in non-FlaggedRevs-enabled namespaces are already filtered out of
the actual list of pages displayed at Special:PendingChanges (and, as
best I can tell from scanning over some past commits, likely have been
since at least bd9519ee3f8d in 2009); so AFAICS this change just matches
the behaviour of the (new-ish) pending-changes-counter to the
long-standing behaviour of the page-list itself.

This patch introduces a JOIN with the `page` table on the SQL query that
fetches the count of pages with a pending change, so that it can also
then filter out any pages for which the `page_namespace` is not a
FlaggedRevs-enabled namespace (as determined by
`FlaggedRevs::getReviewNamespaces()`).

I also changed this query to use `newSelectQueryBuilder` instead of
`selectRowCount`, since the docs for the latter function say that new
callers should use the former function instead [0].

I would have added a test for this (fixed) behaviour....however, the
file in question (PendingChangesPager.php) currently appears to be
completely untested [1]. And I'm not sure if I currently trust myself to
be able to write a FlaggedRevs test file from scratch.

This commit follows up ff7b19814516 (I think).

[0] https://github.com/wikimedia/mediawiki/blob/850b7139907f9b6cec3611345e442f576b0064f3/includes/libs/Rdbms/Database/IReadableDatabase.php#L509-L510
[1] https://doc.wikimedia.org/cover-extensions/FlaggedRevs/includes/frontend/specialpages/reports/PendingChangesPager.php.html

Bug: T413016
Change-Id: I329f9700c04130da46ae13fff9f68e3fea7ee020